+++ DfE.FindInformationAcademiesTrusts.Data/Repositories/Trust/ITrustRepository.cs
@@ -4,4 +4,5 @@ public interface ITrustRepository
 {
     Task<TrustSummary?> GetTrustSummaryAsync(string uid);
     Task<TrustDetails> GetTrustDetailsAsync(string uid);
+    Task<TrustGovernance> GetTrustGovernanceAsync(string uid);
 }