+++ DfE.FindInformationAcademiesTrusts.Data/Repositories/Trust/TrustGovernance.cs
@@ -1,0 +1,7 @@
+﻿namespace DfE.FindInformationAcademiesTrusts.Data.Repositories.Trust;
+
+public record TrustGovernance(
+    Governor[] TrustLeadership,
+    Governor[] Members,
+    Governor[] Trustees,
+    Governor[] HistoricMembers);