+++ DfE.FindInformationAcademiesTrusts/Services/Trust/TrustService.cs
@@ -8,6 +8,7 @@ public interface ITrustService
 {
     Task<TrustSummaryServiceModel?> GetTrustSummaryAsync(string uid);
     Task<TrustDetailsServiceModel> GetTrustDetailsAsync(string uid);
+    Task<TrustGovernanceServiceModel> GetTrustGovernanceAsync(string uid);
 }
 
 public class TrustService(
@@ -61,5 +62,12 @@ public class TrustService(
         );
 
         return trustDetailsDto;
+    }
+
+    public async Task<TrustGovernanceServiceModel> GetTrustGovernanceAsync(string uid)
+    {
+        var (trustLeadership, members, trustees, historicMembers) = await trustRepository.GetTrustGovernanceAsync(uid);
+        var governanceDto = new TrustGovernanceServiceModel(trustLeadership, members, trustees, historicMembers);
+        return governanceDto;
     }
 }