+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Feature flags added to allow for selectively showing features in different environments
 - Add search autocomplete to the header of the application
+- Add Trust Governance Page
 
 ### Changed